Added separate header file for asynchronous logging and updated examples and readme
