+++ CPPLOGGER_SYNC.h
@@ -5,8 +5,8 @@
  *      Author: sajeeb
  */
 
-#ifndef CPPLOGGER_H_
-#define CPPLOGGER_H_
+#ifndef CPPLOGGER_SYNC_H_
+#define CPPLOGGER_SYNC_H_
 
 #include <iostream>
 #include <stdlib.h>
@@ -42,6 +42,7 @@ using std::mutex;
 /**
  * macro to initialize the extern variables in logger
  */
+#ifndef logger_init
 #define logger_init() \
 namespace logger { \
 	bool _print_timestamps_ = false; \
@@ -54,9 +55,14 @@ namespace logger { \
 	mutex critical_section; \
 	va_list __args__; \
 	FILE* _output_stream_ = stdout; \
-	lock_free_queue queue;\
 	timestamp_resolution _resolution = logger::timestamp_resolution::millisecond;\
 }
+#endif
+
+#ifndef logger_output_stream
+#define logger_output_stream(v) \
+		logger::_output_stream_ = v
+#endif
 
 namespace logger {
 
@@ -110,17 +116,9 @@ namespace logger {
 	#define logger_enable(v) logger::_enable_ = v
 
 	/**
-	 *
-	 */
-	extern lock_free_queue queue;
-
-	/**
 	 * The output stream of the logger.
 	 */
 	extern FILE* _output_stream_;
-	#define logger_output_stream(v) \
-			logger::_output_stream_ = v;\
-			logger::queue.set_output_stream((FILE*)v)
 
 	/**
 	 *
@@ -307,48 +305,9 @@ namespace logger {
 	}
 
 	/**
-	 * wrapper function containing the boilerplate code for multi-threaded
-	 * logging
-	 */
-	#define __CPPLOGGER_ASYNC__(fmt, color, type) {\
-		va_list __args__; \
-		va_start(__args__, fmt); \
-		lock_free_queue::node* _node = queue.get();\
-		char* b = _node->data;\
-		if (_print_log_type_) { \
-			b += sprintf(b, "[" color type ANSI_RESET "]"); \
-		} \
-		if (_print_timestamps_) { \
-			b += __print_timestamp_sprintf__(b);\
-		} \
-		if (_print_thread_id_) { \
-			std::stringstream ss; ss << std::this_thread::get_id(); \
-			const unsigned long long int id = std::stoull(ss.str()); \
-			b += sprintf(b, "[%04lld]", id); \
-		} \
-		if (_print_log_type_ || _print_timestamps_ || _print_thread_id_) { \
-			b += sprintf(b, ": "); \
-		} \
-		if (_print_file_) { \
-			b += sprintf(b, "%s:", _file_); \
-		} \
-		if (_print_line_) { \
-			b += sprintf(b, "%d:", line); \
-		} \
-		if (_print_file_ || _print_line_) { \
-			b += sprintf(b, ": "); \
-		} \
-		b += vsprintf(b, fmt, __args__); \
-		b += sprintf(b, ANSI_RESET "\n"); \
-		va_end(__args__);\
-		_node->done = true;\
-	}
-
-	/**
 	 * Variadic argument function for printing information logs to screen.
 	 */
 	#define logger_info(...) logger::_info_(__FILE__, __LINE__, __VA_ARGS__)
-	#define logger_async_info(...) logger::_info_async_(__FILE__, __LINE__, __VA_ARGS__)
 	#define logger_info_mt(...) logger::_info_mt_(__FILE__, __LINE__, __VA_ARGS__)
 
 	inline void _info_(const char* _file_, const int line, const int cvl, const int avl, const char* fmt, ...) {
@@ -378,23 +337,10 @@ namespace logger {
 		}
 	}
 
-	inline void _info_async_(const char* _file_, const int line, const int cvl, const int avl, const char* fmt, ...) {
-		if (_enable_ && cvl >= avl) {
-			__CPPLOGGER_ASYNC__(fmt, ANSI_GREEN, "INFO");
-		}
-	}
-
-	inline void _info_async_(const char* _file_, const int line, const char* fmt, ...) {
-		if (_enable_) {
-			__CPPLOGGER_ASYNC__(fmt, ANSI_GREEN, "INFO");
-		}
-	}
-
 	/**
 	 * Variadic argument function for printing error logs to screen.
 	 */
 	#define logger_error(...) logger::_error_(__FILE__, __LINE__, __VA_ARGS__)
-	#define logger_async_error(...) logger::_error_async_(__FILE__, __LINE__, __VA_ARGS__)
 	#define logger_error_mt(...) logger::_error_mt_(__FILE__, __LINE__, __VA_ARGS__)
 
 	inline void _error_(const char* _file_, const int line, const int cvl, const int avl, const char* fmt, ...) {
@@ -423,23 +369,10 @@ namespace logger {
 		}
 	}
 
-	inline void _error_async_(const char* _file_, const int line, const int cvl, const int avl, const char* fmt, ...) {
-		if (_enable_ && cvl >= avl) {
-			__CPPLOGGER_ASYNC__(fmt, ANSI_RED, " ERR");
-		}
-	}
-
-	inline void _error_async_(const char* _file_, const int line, const char* fmt, ...) {
-		if (_enable_) {
-			__CPPLOGGER_ASYNC__(fmt, ANSI_RED, " ERR");
-		}
-	}
-
 	/**
 	 * Variadic argument function for printing warning logs to screen.
 	 */
 	#define logger_warning(...) logger::_warning_(__FILE__, __LINE__, __VA_ARGS__)
-	#define logger_async_warning(...) logger::_warning_async_(__FILE__, __LINE__, __VA_ARGS__)
 	#define logger_warning_mt(...) logger::_warning_mt_(__FILE__, __LINE__, __VA_ARGS__)
 
 	inline void _warning_(const char* _file_, const int line, const int cvl, const int avl, const char* fmt, ...) {
@@ -466,18 +399,6 @@ namespace logger {
 	inline void _warning_mt_(const char* _file_, const int line, const char* fmt, ...) {
 		if (_enable_) {
 			__CPPLOGGER_MT__(ANSI_BLUE, "WARN");
-		}
-	}
-
-	inline void _warning_async_(const char* _file_, const int line, const int cvl, const int avl, const char* fmt, ...) {
-		if (_enable_ && cvl >= avl) {
-			__CPPLOGGER_ASYNC__(fmt, ANSI_BLUE, "WARN");
-		}
-	}
-
-	inline void _warning_async_(const char* _file_, const int line, const char* fmt, ...) {
-		if (_enable_) {
-			__CPPLOGGER_ASYNC__(fmt, ANSI_BLUE, "WARN");
 		}
 	}
 
@@ -511,4 +432,4 @@ namespace logger {
 
 }
 
-#endif /* CPPLOGGER_H_ */
+#endif /* CPPLOGGER_SYNC_H_ */
