+++ examples/typographic_example.cpp
@@ -5,7 +5,7 @@
  *      Author: sajeeb
  */
 
-#include <CPPLOGGER.h>
+#include "../CPPLOGGER_SYNC.h"
 
 /**
  * The following should always be called only in the main translation unit.
